Refine Next.js intl configuration
