+++ apps/web/src/i18n/index.ts
@@ -1,0 +1,2 @@
+export { default } from './request';
+export * from './request';
