+++ apps/web/src/messages/ar.json
@@ -1,4 +1,6 @@
 {
+  "greeting": "مرحبًا",
+  "welcome": "أهلاً بكم في معهد الفوز للقرآن",
   "common": {
     "loading": "جاري التحميل...",
     "error": "خطأ",
